LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/mixed-content/gen/sharedworker-module-data.meta/opt-in/xhr.https-expected.txt


PASS Mixed-Content: Expects allowed for xhr to same-https origin and no-redirect redirection from https context.
PASS Mixed-Content: Expects blocked for xhr to cross-http origin and no-redirect redirection from https context.
FAIL Mixed-Content: Expects blocked for xhr to same-http origin and no-redirect redirection from https context. assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
